+++ IAR_EW6.7/uip/timer.h
@@ -62,6 +62,7 @@
 #define __TIMER_H__
 
 #include "clock.h"
+#include "types.h"
 
 /**
  * A timer.
@@ -79,7 +80,7 @@ struct timer {
 void timer_set(struct timer *t, clock_time_t interval);
 void timer_reset(struct timer *t);
 void timer_restart(struct timer *t);
-int timer_expired(struct timer *t);
+u8 timer_expired(struct timer *t);
 
 #endif /* __TIMER_H__ */
